+++ src/Trarizon.Library/Collections/CollectionBuilders.cs
@@ -1,11 +1,13 @@
-﻿using Trarizon.Library.Collections.AllocOpt;
+﻿using System.Diagnostics.CodeAnalysis;
+using Trarizon.Library.Collections.AllocOpt;
 using Trarizon.Library.Collections.Generic;
 using Trarizon.Library.Collections.Specialized;
 
 namespace Trarizon.Library.Collections;
 public static class CollectionBuilders
 {
-    public static AllocOptList<T> CreateAllocOptList<T>(ReadOnlySpan<T> values)
+    [Experimental(TraThrow.ExpNoUse)]
+    internal static AllocOptList<T> CreateAllocOptList<T>(ReadOnlySpan<T> values)
     {
         var list = new AllocOptList<T>();
         list.AddRange(values);