+++ src/Trarizon.Library/Collections/TraCollection/_.cs
@@ -5,9 +5,6 @@ using Trarizon.Library.Collections.StackAlloc;
 namespace Trarizon.Library.Collections;
 public static partial class TraCollection
 {
-    public static ReversedSpan<T> AsSpan<T>(Stack<T> stack)
-        => Utils<T>.GetUnderlyingArray(stack).AsSpan(..stack.Count).AsReversed();
-
     #region NetStd_Fix
 
 #if NETSTANDARD2_0